+++ fr.cnes.analysis.tools.export.xml/src/fr/cnes/analysis/tools/export/xml/ExportXml.java
@@ -152,14 +152,20 @@ public class ExportXml implements IExport {
             resultAttributes.add(new Attribute("resultLine", checkResult.getLine().toString()));
 
             /*
-             * The location and message are defined in violations by only one
-             * attribute (String) made this way: [Location -> Message], so we
-             * split it to get the two exploitable strings.
+             * In case no location is defined, it means the violations or the
+             * metric is intended to a file. This way, there is no
+             * resultNamePlace but the resultTypePlace will be "class".
              */
-            resultAttributes.add(new Attribute("resultNamePlace", checkResult.getLocation()));
+            if (checkResult.getLocation() == null || checkResult.getLocation().isEmpty()) {
+                resultAttributes.add(new Attribute("resultTypePlace", "class"));
+            } else {
+                resultAttributes.add(new Attribute("resultTypePlace", "method"));
+                resultAttributes.add(new Attribute("resultNamePlace", checkResult.getLocation()));
+            }
             /*
              * If the analysis checkresult has a result, we add it.
              */
+
             if (checkResult.getValue() != null) {
                 resultAttributes
                         .add(new Attribute("resultValue", checkResult.getValue().toString()));
