+++ fr.cnes.analysis.tools.fortran77.metrics/lex/F77METComplexitySimplified.lex
@@ -115,7 +115,7 @@ STRING		 = \'[^\']*\' | \"[^\"]*\"
 %}
 
 %eofval{
-	this.computeMetric("FILE", Float.NaN, 0);
+	this.computeMetric(null, Float.NaN, 0);
 	return getCheckResults();
 %eofval}
 %%
