+++ fr.cnes.analysis.tools.fortran77.metrics/lex/F77METLineOfCode.lex
@@ -105,7 +105,7 @@ SPACE		 = [\ \r\f\t]
 
 /* At the end of analysis, atEOF is set at true. This is not meant to be modified. */
 %eofval{
-	this.computeMetric("FILE", numTotal, 0);
+	this.computeMetric(null, numTotal, 0);
 	return getCheckResults();
 %eofval}
 %%
