+++ fr.cnes.analysis.tools.fortran77.metrics/lex/F77METNesting.lex
@@ -128,7 +128,7 @@ SMBL		 = \&   	  | \+			| \$
 %}
 
 %eofval{
-	this.computeMetric("FILE", Float.NaN, 0);
+	this.computeMetric(null, Float.NaN, 0);
 	return getCheckResults();
 %eofval}
 %%
