+++ fr.cnes.analysis.tools.fortran90.metrics/lex/F90METNesting.lex
@@ -115,7 +115,7 @@ STRING		 = \'[^\']*\' | \"[^\"]*\"
 %}
 
 %eofval{
-	this.computeMetric("FILE", Float.NaN, 0);
+	this.computeMetric(null, Float.NaN, 0);
 	return getCheckResults();
 %eofval}
 %%
