+++ fr.cnes.analysis.tools.fortran90.metrics/lex/F90METRatioComment.lex
@@ -107,7 +107,7 @@ END			 = END		  | end
 
 /* At the end of analysis, atEOF is set at true. This is not meant to be modified. */
 %eofval{
-	this.computeMetric("FILE", Float.NaN, 0);
+	this.computeMetric(null, Float.NaN, 0);
 	return getCheckResults();
 %eofval}
