--- fr.cnes.analysis.tools.ui/src/fr/cnes/analysis/tools/ui/wizard/export/metric/xml/package-info.java
@@ -1,13 +1,0 @@
-/************************************************************************************************/
-/* i-Code CNES is a static code analyzer.                                                       */
-/* This software is a free software, under the terms of the Eclipse Public License version 1.0. */
-/* http://www.eclipse.org/legal/epl-v10.html                                                    */
-/************************************************************************************************/
-/**
- * @version 2.0
- * @since 2017-07-12
- * 
- *        Contains classes necessary to export metrics view data in the
- *        format XML.
- */
-package fr.cnes.analysis.tools.ui.wizard.export.metric.xml;
