+++ fr.cnes.analysis.tools.ui/uml/wizard_arch.ucls
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <class-diagram version="1.2.0" icons="true" automaticImage="PNG" always-add-relationships="false" generalizations="true" 
   realizations="true" associations="true" dependencies="false" nesting-relationships="true" router="FAN">  
-  <class id="1" language="java" name="fr.cnes.analysis.tools.ui.wizard.export.rules.RuleExportWizardPage" 
+  <class id="1" language="java" name="fr.cnes.analysis.tools.ui.wizard.export.CheckerExportWizardPage" 
     project="fr.cnes.analysis.tools.ui" 
-    file="/fr.cnes.analysis.tools.ui/src/fr/cnes/analysis/tools/ui/wizard/export/rules/RuleExportWizardPage.java" 
+    file="/fr.cnes.analysis.tools.ui/src/fr/cnes/analysis/tools/ui/wizard/export/CheckerExportWizardPage.java" 
     binary="false" corner="BOTTOM_RIGHT">    
     <position height="207" width="191" x="646" y="248"/>    
     <display autosize="true" stereotype="true" package="true" initial-value="false" signature="true" 
@@ -12,9 +12,9 @@
       <operations public="true" package="true" protected="true" private="true" static="true"/>    
     </display>  
   </class>  
-  <class id="2" language="java" name="fr.cnes.analysis.tools.ui.wizard.export.rules.RuleExportWizard" 
+  <class id="2" language="java" name="fr.cnes.analysis.tools.ui.wizard.export.CheckerExportWizard" 
     project="fr.cnes.analysis.tools.ui" 
-    file="/fr.cnes.analysis.tools.ui/src/fr/cnes/analysis/tools/ui/wizard/export/rules/RuleExportWizard.java" 
+    file="/fr.cnes.analysis.tools.ui/src/fr/cnes/analysis/tools/ui/wizard/export/CheckerExportWizard.java" 
     binary="false" corner="BOTTOM_RIGHT">    
     <position height="171" width="227" x="343" y="37"/>    
     <display autosize="true" stereotype="true" package="true" initial-value="false" signature="true" 
@@ -23,50 +23,16 @@
       <operations public="true" package="true" protected="true" private="true" static="true"/>    
     </display>  
   </class>  
-  <class id="3" language="java" name="fr.cnes.analysis.tools.ui.wizard.export.rules.xml.RuleXMLExportWizardPage" 
-    project="fr.cnes.analysis.tools.ui" 
-    file="/fr.cnes.analysis.tools.ui/src/fr/cnes/analysis/tools/ui/wizard/export/rules/xml/RuleXMLExportWizardPage.java" 
-    binary="false" corner="BOTTOM_RIGHT">    
-    <position height="171" width="262" x="41" y="248"/>    
-    <display autosize="true" stereotype="true" package="true" initial-value="false" signature="true" 
-      sort-features="false" accessors="true" visibility="true">      
-      <attributes public="true" package="true" protected="true" private="true" static="true"/>      
-      <operations public="true" package="true" protected="true" private="true" static="true"/>    
-    </display>  
-  </class>  
-  <class id="4" language="java" name="fr.cnes.analysis.tools.ui.wizard.export.rules.csv.RuleCSVExportWizardPage" 
-    project="fr.cnes.analysis.tools.ui" 
-    file="/fr.cnes.analysis.tools.ui/src/fr/cnes/analysis/tools/ui/wizard/export/rules/csv/RuleCSVExportWizardPage.java" 
-    binary="false" corner="BOTTOM_RIGHT">    
-    <position height="171" width="263" x="343" y="248"/>    
-    <display autosize="true" stereotype="true" package="true" initial-value="false" signature="true" 
-      sort-features="false" accessors="true" visibility="true">      
-      <attributes public="true" package="true" protected="true" private="true" static="true"/>      
-      <operations public="true" package="true" protected="true" private="true" static="true"/>    
-    </display>  
-  </class>  
-  <association id="5">    
+  <association id="3">    
     <end type="SOURCE" refId="2" navigable="false">      
-      <attribute id="6" name="mainPage"/>      
-      <multiplicity id="7" minimum="0" maximum="1"/>    
+      <attribute id="4" name="mainPage">        
+        <position height="0" width="0" x="0" y="0"/>      
+      </attribute>      
+      <multiplicity id="5" minimum="0" maximum="1">        
+        <position height="0" width="0" x="0" y="0"/>      
+      </multiplicity>    
     </end>    
     <end type="TARGET" refId="1" navigable="true"/>    
-    <display labels="true" multiplicity="true"/>  
-  </association>  
-  <association id="8">    
-    <end type="SOURCE" refId="2" navigable="false">      
-      <attribute id="9" name="exportXML"/>      
-      <multiplicity id="10" minimum="0" maximum="1"/>    
-    </end>    
-    <end type="TARGET" refId="3" navigable="true"/>    
-    <display labels="true" multiplicity="true"/>  
-  </association>  
-  <association id="11">    
-    <end type="SOURCE" refId="2" navigable="false">      
-      <attribute id="12" name="exportCSV"/>      
-      <multiplicity id="13" minimum="0" maximum="1"/>    
-    </end>    
-    <end type="TARGET" refId="4" navigable="true"/>    
     <display labels="true" multiplicity="true"/>  
   </association>  
   <classifier-display autosize="true" stereotype="true" package="true" initial-value="false" signature="true"
