+++ tests/fr.cnes.analysis.tools.fortran77.metrics.test/src/fr/cnes/analysis/tools/fortran77/metrics/ComplexitySimplified/TestF77METComplexitySimplified.java
@@ -49,7 +49,7 @@ public class TestF77METComplexitySimplified {
             final List<CheckResult> checkResults = metric.run();
             CheckResult fileValue = null;
             for (CheckResult check : checkResults) {
-                if (check.getLocation().equals("FILE")) {
+                if (check.getLocation()==null || check.getLocation().isEmpty()) {
                     fileValue = check;
                     checkResults.remove(checkResults.indexOf(check));
                 }
@@ -60,7 +60,7 @@ public class TestF77METComplexitySimplified {
             } else {
                 // CheckResult fileValue;
                 for (CheckResult check : checkResults) {
-                    if (check.getLocation().equals("FILE")) {
+                    if (check.getLocation()==null || check.getLocation().isEmpty()) {
                         fileValue = check;
                         checkResults.remove(checkResults.indexOf(check));
                     }
