+++ tests/fr.cnes.analysis.tools.fortran77.metrics.test/src/fr/cnes/analysis/tools/fortran77/metrics/RatioComment/TestF77METRatioComment.java
@@ -50,7 +50,7 @@ public class TestF77METRatioComment {
             final List<CheckResult> checkResults = metric.run();
             CheckResult fileValue = null;
             for (CheckResult check : checkResults) {
-                if (check.getLocation().equals("FILE")) {
+                if (check.getLocation()==null || check.getLocation().isEmpty()) {
                     fileValue = check;
                     checkResults.remove(checkResults.indexOf(check));
                 }
