+++ tests/fr.cnes.analysis.tools.fortran90.metrics.test/src/fr/cnes/analysis/tools/fortran90/metrics/Nesting/TestF90METNesting.java
@@ -51,7 +51,7 @@ public class TestF90METNesting {
             final List<CheckResult> checkResults = metric.run();
             CheckResult fileValue = null;
             for (CheckResult check : checkResults) {
-                if (check.getLocation().equals("FILE")) {
+                if (check.getLocation()==null || check.getLocation().isEmpty()) {
                     fileValue = check;
                     checkResults.remove(checkResults.indexOf(check));
                 }
