+++ tests/fr.cnes.analysis.tools.fortran90.metrics.test/src/fr/cnes/analysis/tools/fortran90/metrics/RatioComment/TestF90METRatioComment.java
@@ -49,7 +49,7 @@ public class TestF90METRatioComment {
             final List<CheckResult> checkResults = metric.run();
             CheckResult fileValue = null;
             for (CheckResult check : checkResults) {
-                if (check.getLocation().equals("FILE")) {
+                if (check.getLocation()==null || check.getLocation().isEmpty()) {
                     fileValue = check;
                     checkResults.remove(checkResults.indexOf(check));
                 }
